The actual file for reading RDFs
Functions were written a bit differently, comparing to LAMPSQw, in order to make sure the proper array memory allocation as well as closing files which are used for reading the data. Function parse_dep_var will need an input-file as well as the number of RDFs to read from that file.